lib/modelporter/pyproject.toml
[project]
name = "modelporter"
version = "0.1.0"
description = "A play on porter (one who carries things) and exporter for models"
authors = [
    { name = "Christopher Mortimer", email = "christopher@mortimer.xyz" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-aiplatform = "^1.72.0"
hydra-core = "^1.3.2"
rich = "^13.9.4"
click = "^8.1.8"
docker = "^7.1.0"
conman = { path = "../conman", develop = true }
google-cloud-run = "^0.10.14"
google-cloud-build = "^3.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[tool.poetry.scripts]
modelporter = "modelporter.cli:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
